Laying out some testing ideas for using xml files with java
Want to go into playing around with Java with .xml files to see how to start storing the data from the from game and get a feel for ideas of how to organize the data.
Then going into take that data and add them around and play with how to store will be next one.