Setting up on new device
So setting up on new device and found a error when trying to compile the code because of the external .jar files that where downloaded could not find the path to where I stored them. So for now I put them in the newly created "lib" folder in order to help fix further problems. EK